fix:Notification crash on NotificationUtils.kt * Changed the code to pass reminderItem object using safe args. * added global action in FragmentReminderDetails.kt

+++ starter/app/src/main/java/com/udacity/project4/view/selectreminderlocation/SelectLocationFragment.kt
@@ -58,8 +58,8 @@ class SelectLocationFragment : Fragment(), OnMapReadyCallback {
                 val action =
                     SelectLocationFragmentDirections
                         .actionSelectLocationFragmentToSaveReminderFragment(
-                            latitude?.toInt()!!,
-                            longitude?.toInt()!!,
+                            latitude?.toFloat()!!,
+                            longitude?.toFloat()!!,
                             locationName
                         )
                 findNavController().navigate(action)
